OneDrive/Desktop/my meds new/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
requests
python-dotenv
